A bit of work on the toy TUI editor sandbox
I need to develop the "theory" behind how layout (and GUI in general) is
going to be done etc. Hence this. If all is good one'd be able to
substitute chars for pixels and add a measurement fn or two to get this
going with a proper UI. Also a good benchmark. Currently runs poorly (as
expected).