Implement the sketch optimization via SketchSubset operator
Reduces the time it takes to run tests for me by about 300 ms. This is a
probabilistic opt so sometimes it works too well and sometimes it does
not work at all. Followed up by boundsand depth checks soon to come, the
situation should improve a bit more (but would it?)