Basic implementation of the bounds optimization
Doesn't improve the timings too much. We're still about 1s for a run of
tests on my machine. My rough goal was 100ms-ish? Maybe because the
editor relies more on depth rather than bounds, will see. Despite no
significant improvement we still manage to reject hundreds of thousands
of mismatches before they're passed on to ahead with Boudns, which is an
OK result. Combined with an outer `%sketch`, we are able to reject
millions of mismatches before ahead.