Upload progress in O2 opt, minor fixes, renames
Rename some operators for consistency (huh?)
Rename `Term::Dict#pairs` to `Term::Dict#pairspart`, add
`Term::Dict#itemspart` as well.

Progress in O2: I'm now seeing the tests run in ~700-800ms, still takes
a lot longer that it should. Previously was 1s-1.2s, the improvement is
obvious, but still, some more optimization is required. My original goal
was <100ms, not sure if I'll be able to hit it. I do not have a lot of
tricks left but will see.