More O2 optimization efforts
Now we're somewhere around 700ms on my machine. With pure rewriting,
1100ms -> 700ms is an OK reduction. There are some more tricks regarding
pattern partitioning etc, will see if they help to get <500ms. But
again, with the goal of <100ms 500ms is a lot! Note that I'm not done
with the engine itself, this is just compilation. Good that it runs so
fast (huh?) without me touching the brains, just phrasing differently.