Get rid of all uses of `ItemAhead::Fn`, remove `ItemAhead::Fn`
Also: force `%many` and `%group` to behave similarly in that an explicit
pluralization is required in backmaps, i.e. `(%many xs ...) <> {xs:
...}` would replace with one, but `... <> {(xs): (...)}` would replace
with many. Not sure that's the right move, but "explicit is better than
implicit", nah? A drop in the ocean, I know, but still. Wrote a few
failing tests, fail due to `SuccessorsView` and `neighbor?` crap on
`Operator::Item`. Will see what I can do about them. Failing tests are
rare these days. Very happy to see some. Currently we're about 600ms
(100ms-ish drop from the Fn removals), but the failing tests may obscure
a few ms here and there, and in general there is imprecision in the
measurements. Will see.