Implement `KeypathQuery` instead of `KeypathTip`
I'm not sure about the performance. Seeing more regular dips into <600ms
after this but still, I'd say it's mostly unaffected. But that's on
tests. Benchmarks of individual patterns show a few microseconds of
improvement. But the keypath side of `match()` is in general slow, we do
not expect it to be fast (not until a proper dict implementation, and
even then...) Note that our current dict implementation is highly
dependent on the hash function which makes the benchmark numbers jump
all over the place, since here we're allocating Item/PairNodes and there
we aren't. Implementing a linear probing Item/PairNode items array
should supposedly solve this. Also the whole backmap thing is so slow
it'd probably overwhelm any bit of improvement. Will see.