Work on the parser, implement some shorthands & remove remains of the old ones
Most shorthands are in the pairspart and accessed through `¦`. If you
don't use `¦` you won't get them, and that's by design. `x_: ...` is a
shorthand for `x: x←...`, `x⋮ ...` is a shorthand for `x: (%optional ...
x_<type of ...>)`, `-x_` is a shorthand for `x: (%- _ x)`, `{_ ...}` is
a shorthand for `(_* ¦ _ ...)`, `x_` inline is a shorthand for `x: x_`
e.g. `{_ x_number y_number}` is the same as `(_* ¦ _ x: x_number y:
y_number)`