Refactor keypath: Tip and Label are the same thing
KeypathQuery is now Keypath::Appender
Its tip and BackmapTrie's Label are now the same object, Keypath::Word.

The "phrase"/"sentence"/"word" analogy is an attempt to build an image
in my mind that helps me understand what is going on. Won't say it is
failing but ... :^) It's all pretty confusing.