`Item.match`: begin refactoring, group ord, behind into Progress
It's so scary there, gosh. Hopefully will refactor into something nicer
because right now I barely understand what's going on there. This isn't
best news since we have a failing test and the reason is somewhere in
there. Or rather, in that I have to figure out a rule that will resolve
the problem in the test, and then implement it somehow, and that's the
scary part looking at that code.