Fix the only failing test! And write some more for similar scenarios!
Now even distribution can be achieved predictably. Which is nice! What
we do overall is for extremely ambiguous patterns to have *some*,
semi-predictable handling. Such patterns are likely *not* to appear in
practice, therefore, we do not pay significant attention to the very
fact of "SEMI" predictability; they are handled, but being extremely
ambiguous, can depend on content etc. -- the client's responsibility if
they decide to write these kinds of patterns.