editor: fix `move-parent-below`/above bug in toplevel dict
We do not want to do `(X): ...` in toplevel dict because we don't know
who is going to consume it -- will they adhere to our `Rewrite.many` or
will they not, wrapping us in a dictionary every time we do a rewrite?
So introduce a further required level of nesting and bump up editR
relR's ascent to 3.