d7: Refactor `transform` node
Currently this requires the introduction of a companion to "cycle",
"initialize"; which slows everything down by some amount. With the
upcoming arrival of suggestions for everything the number of patterns in
nodestep will increase, by a lot. New nodes will contribute as well.
Making the `Term.case` in there a very very bad choice, due to its
inevitable O(N) nature. However the general design of D7 is bad right
now (it's a prototype), and optimizing nodestep as an idea isn't the
easiest thing in the world. I think there are more low-hanging fruit
before we tackle nodestep (a required affair if we want decent runtime
perf). Also, the very idea of "initialize" must be reassessed and
defined more clearly -- when do we emit "initialize"? I suppose we
should emit it when new nodes are added -- but how do we detect that
efficiently, without resorting to population? (the idea of "initialize"
is to avoid getting into the population)