Fix a very old bug
This code is old... I would really like to remove autocasting at some
point. It was a mistake. But it is so engrained into the source code
it'd be really hard to remove it. It is viral in a bad sense. Crystal
cannot produce backtraces when the method is missing; and also since it
is bidirectional (upcast downcasts and downcast upcasts to find the
method), if the method does not exist, it causes an infinite loop --
sometimes, not always. Neither is good DX.