Try to offset suggestions into visibility before clipping
The calculations all feel extremely wrong but in most cases this appears
to work. The vis code would probably be ditched regardless since there's
just too much wrong-ness going on -- for one, we need an intermediate
tree, and soma5 seems like a good candidate once ported to TUI. So this
solution will work for now.