d7, nitrene: Basic async behavior for `transform`
Not sure how logically sane everything is but it appears to work. With
tests, it is unclear how to manage transforms that now may not
necessarily finish in time. Right now everything works and will probably
work regardless because we have pretty big bounds on d7 tests -- and
they do not terminate the way e.g. soma does, they only terminate when
they reach the goal or when their limit is exceeded. The latter can
simply be bumped up in case transforms do not appear to complete in
time.