pprint2_vis: Add basic save/load and some examples
The thing is starting to fall apart. I want to have a command palette
but it's not going to be possible with pprint2_vis. Similarly the whole
`@events` approach with Nitrene and Termbox is flawed, and probably
doesn't work in all cases. A different event management technique is
needed. Maybe via AtomicSleepyQueue or something along those lines.