tspace: upload surf2 progress
We are able to do some over-the-network stuff but everything feels so
brittle! Still no notification mechanism -- that's the nearest TODO; the
second nearest TODO is incorporating changes into some kind of view that
Tconn provides to sink (currently we're simply logging at that point)
and not calling sink. This will allow us to run tests (hopefully!) After
tests we're supposed to connect the whole thing to D7. Also maybe
simulating a distributed map backend (or even using a real one; we need
a distributed map only for routing, i.e. given this key digest give me
the corresponding peer IPs; then it's the usual (even the current RPC
perhaps) protocol)