soma, d7, ufold, uiR: move to relative coords for left/top
- We now use parent-relative dl/dt instead of absolute l/t.
- This is a small first step toward making uiR cache-friendly. Before,
even the smallest change would have caused invalidation of lots of
things; now there's at least a chance for things to be kept in cache for
several frames at least.
- There are issues with uiR itself right now that prevent good caching
behavior. I'm not a fan of changing uiR because I like the way it is
organized right now; however that is very cache unfriendly. More
attention toward the order of rewrite events must be paid & the rewriter
circuit reorganized to ensure better cache behavior.
- Reduce drawing in `Reducers.microfold` & related