soma, d7: work on fixing the tspace feedback loop
The bug was pretty rare and requires to wait a few mins before it
appears -- feedback loop settles when it shouldn't. The error appears to
have been in `Mailbox::State#settled` where we must settle only if the
queue is empty, atomically; and otherwise restart the mainloop on the
doc thread. Otherwise there was a state "hole" where we stopped the
mainloop but did not spawn a new doc thread, if preempted just right
way.