soma: Add support for providing const cell caption to `button`
Currently this f***s up the latter because it has an ID assigned at
annotate and that becomes visible. I'll be fixing this shortly with a
brand new & much more cooperative way to implement clickability, hover,
and activity. Or, uhmmm... hopefully it'll end up working, yeah. Less
advertisement, more work.