patterns: Add `Entry::Present`, sort entries in `%layer` by cost
Especially the now-guaranteed Absence checking done first is an
important update. We also do Presence checks just after absence checks
(absence checks are usually "did I already do this" flags so we give way
to them). Keypath-related nodes are considered "expensive" and are done
last. This doesn't improve performance by a huge lot but gives a bit of
guarantees -- if the patterns are expensive one can add an e.g. absence
check of some kind and have guarantees that it'll be checked first.