d7, soma: Move `view` node entirely to instantiation
We now have two separate nodes, `alloy` to do templating and `view` to
show the actual template. We're missing another node that combines
`cell` and `view`, but that's not as urgent right now. And perhaps I'll
be able to come up with a way to actually implement reusable
"components" in D7/Rhodium. Currently it's impossible; I have to
implement nodes manually & lots of functionality gets duplicated &
messy. Implementing most nodes is very easy and it is similarly easy to
build things up from that point; but currently there's no way to "reuse"
this "building up", besides pure copy-paste.