tspace: some removals, refactors & more experimentation
- Remove referrer refcounts / INCREF/DECREF semantics. Now referrer is
"reference id" and that's it. No further refcounts etc.
- Introduce ISet along IMap for places where we had identity values
- Refactor multimap implementation into a separate module
- Attempt to implement IMap on top of ISet. Currently failed. Thick
thick brick wall on `ref` there, its put if absent semantics is very
hard to do on top of a multimap -- the only interesting "emergent" data
structure we can implement on top of a degrading & lying `Set` with
hopes to get something proper out of it. Not sure if it's possible. But
will try to come up with something. The end-product is just too awesome.
Emergence in its glory. Hashes flowing around, nobody knows what they
know, and yet here it is -- out of all this activity emerges a Tspace
over which one can talk arbitrarily!