soma: Add concealed mode for document, reorganize fonts & add Material Icons
Concealed mode partly solves the issue that we have with single-threaded
SFML now, that we require an "exchange" to take place between the
document and the main thread to do UIR -> dwUIR conversion. This is
obviously capped at screen refresh rate since the main thread has the
ability to do rendezvous with the document thread only at a cycle =
end-of-frame. The user now has the option of pressing `(key escape)`
which I freed up in one of the previous commits, to enter a "concealed"
mode -- rendezvous no longer takes place, the user is "out of the loop",
and the document may run at full speed (which is even now when
everything is super slow, vastly exceeding even 60FPS).