d7: Change the syntax of the `transform` node
For better pretty-printing we're using a different `transform` node
syntax from now on:

  (transform (@pin_ to @pout_) body_)
  (transform (@pin_ pattern_ to @pout_) body_)
  (transform (@pin_ to @pout_ with @state_) body_)
  (transform (@pin_ pattern_ to @pout_ with @state_) body_)

The spec in parens is usually pretty short. We'll introduce a special
layout for cases like these where the symbol `transform` and the spec
are going to be put on a line with "longer" width, and the body is going
to be indented and on the next line. This layout will also help other
transform-likes such as `alloy`.