uiR, ufold: Add `flow-wrap` & underlying `x-stack wrap: true`
We can now x-wrap arbitrary UIR nodes, yay!

This works very very slowly, like, slow as hell! So slowly that I've
elevated the priority of the dict rewrite a bit more in the plans. It
seems the majority of performance problems right now arise from
`Term::Dict`.

I'm not sure I want to do the dict rewrite right now but it's getting
more and more urgent! It's a fairly big rewrite that would take a few
weeks of work -- the entire system depends on dicts being fast now, and
it's "fast" in the smart way, not "handwritten assembly" kind of fast,
not that the latter isn't smart. The latter kind of opt I'm really
really bad at; the former, we'll see. The problem is, I need to collect
& organize all "wish the dict would (index this/give this in constant
time/etc.) for me" of all subsystems (and there are plenty of them!)
into a coherent data structure that's as thin on indirections as
possible.