soma, tspace: introduce Meridium surf4, detach soma from surf2
- Prepare soma for surf4 Meridium.
- Run tests using surf4 Meridium.
- Add some more nasty tspace tests.
- The new implementation, surf4, is running on top of a multiset instead
of a map. This means it is vastly friendlier toward completely
chaotic/emergent environments. It is also pretty secure. Nothing but
SHA256 (perhaps) hashes flow through the network. Now we use the term
`secret` instead of `selector`, since it is now impossible (or very hard
computationally) to figure out something even *is* a secret, or belongs
partly or in whole to a secret of some kind.
- We now speak of "atom cost". The data structures of Meridium currently
do not, but hopefull will be optimized later on for such atom cost to be
reduced. An "atom" is an identity out of which the larger data
structures are built, and which e.g. peers store. Right now atom costs
of sensors and appearances are large, in the thoudsands for tiny sensors
and appearances; in 100s of thousands of atoms for large ones.
- Atom cost in practical (not inmemory) environments must have a "cost"
of some kind. Thus in the future we plan to employ proof of work to make
"minting" of atoms somewhat costly. This way, it will be computationally
expensive to attack the system by flooding it with useless atoms.
- Similarly, `Activation`s should also have a cost, since their delivery
in an unstructured p2p network is rather costly, and many peers must
commit their "time & energy" in doing so.
- Each peer will verify all incoming atoms and activations, rejecting
ones it deemed invalid. This will hopefully prevent an attacker from
flooding the set/system with random atoms and activations.