tspace: Work a bit on the thread-safety of `Tconn`
Still a very rough prototype powered by lots of crapcode. Will see what
I can do about it. It's appears that a Tconn has some sort of a Kernel,
which is protected by a lock. We also lack an observer activation queue.
It is an important element to prevent recursive locking if the observer
happens to reference the Tconn recursively. Will have to think about
this.