tspace: Finalize thread-safety of `Tconn` throughout its interactions
Still not sure about it but seems a bit more stable. The checksum &
constraint stuff works too, in fact, and saves us from incompleteness of
data in the underlying set -- hit sometimes if running in parallel, when
e.g. one thread is removing what another is trying to read. So the
checksum stuff was useful after all!!