tspace: Work on a crude remote client/server combining `ISet` and `IChat`
It kind of works but there is so much needless complexity in there that
I'm actually not sure. I'm also very much unsure about the amounts of
traffic. Yes, we're going to buffer things; right now we do not, and
this creates lots of overhead. But there is still a lot (a lot!) of
traffic per client. With this amount of traffic my "P2P dreams" could be
shattered; I'm not sure if modern networks are capable of handling so
much stuff. Will see though. Maybe buffering & compression help ease the
load a little bit.