tspace: Add buffered set & support flush in Tconn
The thing is getting messier and messier. Bad sign. I think I need to
redesign the system here. Definitely the surface stuff & Tspace should
be split from ISets completely, and moved away. It works fine & I find
it rather beautiful of an idea, despite being so resource-intensive etc.
We can work our way up from there, I guess, to something like the
current Tconn. At the moment lots of things are intertwined, and at this
(moderate, but still) scale I find it harder and harder to work on
something that is so intertwined. Some partitioning needed before we can
continue with progress.