tspace: Start work on surf5
Hopefully this revision will be the final one. I've already managed to
optimize BytesMultimap some, but more work ahead. Using BLAKE3 now for
hash since it appears to be faster (esp. for intermediate digests).
Using recursive hash for BytesMultimap to save on space/digest state
copies. With BytesMultimap done, the majority of Tspace data structures
can be moved into surf5. The endgoal of surf5 is to have a `Sensor` and
an `Appearance` object on which we can call `#each_atom` or something
along the lines; and also `each_complement(membership fn, &)`. This
should help hide away all this complexity and leave us to implementing
proper surface management machinery (what is in surf4 the scary nasty
Tconn).