tspace: Upload progress on surf5
Honestly I like this code less and less. I'm more of a "functional"
person these days and this code is a total and unreadable mess, all of
it; working with it is impossible. More & more in favor of redesigning
it yet again. But I don't know how. Whereas mounting is more or less OK
(but should be rewritten a bit here and there), query methods aren't. I
dislike that one must pass a callback of some kind. It would be nice if
we were returning something instead. But what? Also, the thing is quite
parallelizable -- some queries can be built & run in parallel, as well
as mounts. But again, with mounts it is somewhat clear how to do that;
what about queries? The idea of plans comes to my mind -- we should
return plans from queries, which is certainly much more pure than what
we have right now. But would that *really* solve anything?