tspace: Work on surf6
Decided to start a new cleaner impl based on what I've learned from the
previous ones. Seems to be the best one so far. Able to leverage
parallelism in some places (and it actually works!! I mean, sometimes;
i.e. we get 2x to 4x improvement as expected on a 4-core machine). There
are still hard-to-parallelize areas (mainly Xgraph). But still. Will
leave it sequ for now. Only Utrie and AppearanceRegistry (what is
AppearanceField in surf5) left to do.