tspace: Work more on surf6, AppearanceRegistry practically done
Yes all of this is alien technology, more or less; but it is also very
very hard to debug!! So I'm not sure whether the stuff I'm writing
actually works! Especially will all that multithreading stuff -- but
it's necessary; serial is just too slow, esp. on hashing heavy
compartmentalized tasks -- tasks for which mt was actually intended! I'm
avoiding heavy tests because I hate implementation tests; I change the
guts a lot and rewriting impl tests isn't fun! So instead I am shooting
for the higher level tests -- in our case tspace tests at the very top.
But until we get there with surf6, yes, I've no idea whether all of this
even works. My brains ain't too powerful, I see some of it but not all
of it, and keeping the edge cases in mind is especially hard with MT!