tspace: Implement `Sensor` and `Appearance` surfaces in surf6
Yay! At least appearance-by-sensor query is behaving more or less as
expected in terms of bounds! We can index hundreds of thousands of
appearances in a way that makes "exact enough" sensors look them up in
sub millisecond times. And all that running on top of a set! A single
damn set! Some cases degenerate though with app-by-sens, due to an
O(n^2) impl of intersection. Once that is fixed we'll have more or less
predictable bounds on app-by-sens; then it's going to be time to see how
sens-by-app performs. There's also the question on how to parallelize
mount of sensors and appearances. Inserting a million appearances takes
approximately 3gb of RAM (that's a lot! with atom cost = about 2.4mil,
multiply by 32 bytes per atom, the ground for us is ~75mib, say 100mib,
so we're overshooting by quite a lot). It also takes a whopping 3
minutes or so. This is a very very very long time. CPU usage in my case
sits at ~20%. Heavy heavy underutilization. We should be able to spawn a
few workers to speed things up. The only bottleneck is the set and it's
bucketized. Enough of my rambling though. I'm quite happy that this
"alien stuff", especially the set idea, is working at all.