tspace: Improve performance of xsect in surf6 `AppearanceRegistry`
Some queries are inherently slow though, and there's little we can do
about it, at least initially. E.g. the `{color: (0 _ _)}` query I'm
testing completes up to & downloads appearances for both `_`s (in our
case there are hundreds of completions at the ends of which there are
roughly 1/256*1mil appearance ids); and also all appearances with `0`
first value. At least we're not downloading those tens of thousands of
appearances as in the previous impls, and only doing a fairly nasty
intersection. The more precise your query is the better the performance
will be, e.g. with `{x: 0, color: (0 _ _)}` you'd be getting much better
measurements although we'd still be getting the completions for `_`s.
You can simply leave them off but then we won't know if `color` has one
value or two or three. The fact is, at this point we've no idea how
these `_`s are going to be used. So I don't think there's any point in
doing anything with this any further. With some feedback from progress
in the future, maybe we'll be able to improve the situation a bit.