tspace: Add crc16 checksum to WWID
Managed to cram 2 byte checksum into WWID, was very hard!! We're
treading on very dangerous grounds anyway, relying only on 5 byte millis
+ 5 byte randomness for global uniqueness ... But there's little choice.
I tried bumping up to 24 bytes and it's horrible, especially for
appearances. Memory use skyrockets. Anything lower feels weird. As far
as I understand:

- 5 byte millis allows us to count up roughly to 2059, good enough
- 5 byte randomness leaves us at somewhere over 1 million choices --
until theres >50% chance of collision; insofar as I understand the
birthday paradox correctly that is.
- We also have slot which gives us per-conid uniqueness. It's up to the
connection to manage uniqueness of slots; and it can do that with a
simple global atomic counter or something.