soma: Remove `only_settled` from alarm/enqueue, won't work
Caused early document loop termination because it was buggy! It's not
that needed anyway. It's called (from) alarm for a reason. It must be
used thoughtfully. Inquiring about whether a document is settled or not
is a bit more complicated & involves more intermediate states. Too much
price to pay for nothing in particular.