tspace: Move all of Axis to src/
Now that I know how we can interact with a remote termspace, it's time
to rewrite Conn & Node again!! Unfortunately I dislike the current
implementation very much, and it has several outstanding bugs, the most
important of which is effect accumulation if we're offline. Conn & Node
should be delta-based not action-array-based. I.e. we should make Node
immutable, have @syncd node and @staging node, and do a delta which will
yield some effects. Additionally we must be able to ask Conns/Nodes
about which surfaces were sync'd and which ones weren't.