term: Use `Dict#each_entry_ord` where ordering guarantees are required
I thought the hash function would help with ordering entries but I was
just being stupid. It would help, sure, but only until collisions.
Whenever we have collisions, in the current dict impl (& in the future
one) we'll depend on insertion order due to how collisions are handled,
which means we loose global order stability. Similarly with the way we
implement symbols, there is order dependence there as well.

This commit implements a "fix" for that -- we simply use a more
expensive, ordered variant everywhere where we "leak" the raw order of
entries to the outside world. Namely, it's the pretty print algorithm,
`ML.compact` and `ML.display`, `Term::Dict#inspect`, and `(entries
_dict)` primitive. I think that's all of the ways one can observe entry
order "from within" at the moment.

Note that we will still be using FNV1a because we need globally stable &
pure `(hashcode term_)` primitive