meridium: Encrypt activations featuring secret-protected stimuli
Currently the only hard-coded algorithm is the ChaCha20-Poly1305
algorithm from OpenSSL. No idea whether there's any point in adding new
algorithms. I'm by no means a security expert, especially on using
blake3 as a pbkdf (is that what the thing is called??) Anyway. Appears
to be reasonably secure, but obviously we lack cipher negotiation etc.
and simply fail to encrypt & do not send if the alg is unsupported.