meridium: Reorder SurfaceRemoved/SurfaceAdded effect yields
Still not sure if I'm satisfied with the new Conn/Nucleus combo. We are
unable to track complements right now, which means we need a very
"resource-heavy" complement search on removal.

This reorder appears to prevent one symptom of this, namely that sensors
can see the absence of an appearance and then its immediate presence on
update. This would work only occasionally -- since we never guarantee
effect order maps to message order nor that message order exists at all!

Ideally, appearances must be aware of the sensors that perceive them,
and sensors must be aware of the appearances that excite them. Like,
each surface must have literally a list/set of complements, or perhaps
we should represent that as a graph. This graph must form via feedback.
That is, we must have heard a response of some kind from the complement
candidate that we found through each_complement/complement_set -- before
we promote it to our list of complements/graph. Then when we change our
value we must notify our complements about that -- but not about any
kind of "removal", like we do now. We know the appearance persists; we
*are* that appearance; why should we send a "removal", like we do now?!
Similarly when the surface is removed the complements must be made aware
of that. Obviously polling (opposed to purely sensor relook) is
required; ideally appearances poll for sensors and sensors poll for
appearances, rarely; say, every minute or so, randomized -- to detect
crashes of their complements. I'm not sure how much useless traffic this
would generate.

I'm quite tired of this part of the system but it must be done at least
somewhat properly, so I'm thinking of redesigning conn/conn nucleus yet
again. This will still have very little chance of success due to heavy
network traffic -- especially in p2p setting which is the original plan;
Axis is only a temporary solution to have some kind of prototype going.