d7: Improve `#population` tracking at expense of performance
We are now able to track not only identities but also where those
identities are in the document. If we forget for a moment about the
amounts of compute that this requires during transition, this is quite
robust & powerful. I believe this will help to improve the editor
experience in the future as well; for instance, this could power a "jump
to definition" type of thing for cells.

Now invited nodes are able to know about their identity predecessors,
expelled ones -- about their successors, if any; not only that, but also
where those are, just a dict lookup away. So we can now handle some kind
of communication going between e.g. nodes that share identity. We will
also be able to do rewriting similarly -- once the rewrite array patch
is there; that is, by directly modifying nodes which have the identity
of interest.

This is required for the module update; not necessarily the nodepath
tracking part, but the idea of having a multiset for population.