d7: Add support for scopes through `module` nodes
This is a very big and long-awaited feature, evolution-wise; one I
thought was impossible or very hard to do. To this point none of the
prototypes had scopes. This solution is a bit hacky, but appears to
mostly work -- and covers the majority of edge cases. Using UUIDs for
modules may be an overkill but will do for now. There probably are
better ways.

There are plenty of bugs/complications left to resolve after this
commit. Most are consequences of the edge type restriction relaxation;
since edges can now contain the cursor (i.e. the pattern for edge will
match), we are now expected to handle that in D7 and above -- M1 is not
our aid anymore. A probable solution for this is to just add a
cursordepth option that "adopts cursor in edges". I'm not sure I want
the restrictions back since they still make no sense in general.

Some other issues include the lack of `#shadow` on `module` which could
lead to peculiar editing experiences; we're desperately trying to hide
`#id` from the user but the editor is still fully aware of it -- and
thus won't let the user do certain things that they would expect to work
given what's before their eyes -- an annoyance to be fixed.