soma: Basic color coded highlighting for edges
Note that I don't think there is any point in making the highlighting
scope dependent. There are just too few colors (hues, really) to work
with. Even now there's quite a lot of collision -- the hash function
isn't super good at picking colors that are different enough. I'm not
sure if this feature is really necessary, but let's keep it for now, and
see if it's useful or just creates visual noise.