d7: Fix `#population`-related transition bug during doc replace
E.g. for cell count, the population of D0 contains `(cell count)`
identity -- as does the population of D1 which D0 was fully replaced
with. Thus D1 thinks it doesn't need to initialize cell count because it
is initialized already -- wrong!

The fix -- if it even is one! -- is to do population-based stuff only
when **both** D0 and D1 are apparantely tracking their populations. This
is obviously true when transition is driven by time; on the other hand
with replace, there isn't any `#population` there; we do a "cold start".
Note that undo/redo doesn't trigger this since for undo/redo both D0 and
D1 have `#population` -- unless it was a replace of course.