soma: Do not clear hover on inactive buttons
Clearing on inactive buttons was previously done to prevent the user
from seeing the `hover:` prop being used. But now that it is a feature
and any node can have it, there's no *real* point in hiding it. In fact,
it goes "against the main line". Moreover, if one accidentally positions
the cursor in `hover:`'s value, the cursor would be destroyed. This is
especially bad because we do not have a way to respawn it yet.