d7: Allow `exposes: (_*)` on any node instead of `module` node
Any node, especially passable nodes (e.g. group, unit, frag) can be used now to create
scopes; the `module` node was a bit clumsy. And what's emerging here is, I think, there
are "nouns" or "things" -- nodes; and "properties". Scope isn't necessarily a "thing".
It's more of a property; which scope does a thing -- a node -- belong to?