d7: Split `handle` into `handlepp` and `handleip`
- handlepp handles pairspart reaction of the node
- handleip handles itemspart reaction of the node
- move mutator into the pairspart under `self: @pin`. This exposes the
bug with rewrite conflicts better; that was previously more or less
hidden in mailbox/related code.

Pairspart offers a "second dimension" for nodes / node behavior. I have
noticed this with uiR etc. So I think we will be more deliberate now
with how node pairspart is used -- as a "second dimension", a way to
escape itemspart identity of the node.

This commit introduces some failing tests which I'll hopefully fix!