soma: Extract unrelated to separate files
- D7VR is its own thing. It's quite f**ked up right now and causes a lot
of bugs; but will work for the time being; it's not what I want to work
on right now.
- Examples (welcome screen, F2) are now in the examples folder. They're
loaded on boot. Not what should happen; but better than being
hard-coded. Small steps lead to great improvement!!1

The remainder of soma6 is probably going to be rewritten/redesigned in
its entirety -- i.e. we'll have soma7 in some amount of time and would
be running that instead of soma6.