soma: Upload progress on soma7
Nothing interesting yet. A slight change of the pipeline. There are now
more steps. Including the model step (aka component tree in the "wider
world"). Hopefully this model step will help with managing focus & other
complex so-called "interaction states". We'll have a bunch of
rewriters/rewriting logic dedicated to that. I think a mailbox
(inbox+outbox) model should be fine for isolation & it is
rewriter-friendly. I'm in complete darkness though at the moment, so
cannot say for sure :^)