ufold: Remove `min-w-sm` from src
This HIDES but does not FIX the bug with editing the welcome screen. The
implementation of `x/y-expand` is very bad and must be redesigned. I
don't know why I struggle with these nodes instead of redesigning them
to be more convenient. It's as if I'm not the one who implemented/can
implement them.