patterns: Add `(%pipe ml _)` to enable parsing WwML in the patern
Currently this is very "fishy" invoking the entire parser this way. But
later on when I get to rewriting the WwML parser (and documenting the
syntax!!! It's crazy the entire thing is still in my head & in the
poorly readable code here and there) -- when I get to rewriting the WwML
parser I hope we'll get a way to enable/disable individual features,
e.g. only parse fractional numbers, only parse booleans, etc.