ml: Upload bits of progress on the new WwML reader
- The new grammar is still under development & about 4x times slower
than the current ml/text impl. On the other hand it is 4x more
flexible!! There is also man/wwml which is an unfinished attempt to
document WwML. Once the grammar is ready I'd feed that back into
man/wwml, for now bits and pieces of it are out of sync.

- Use the display method of the new escape grammar to display strings.
This should finally fix the crappy string escape sequence rendering that
was there all this time.

- Minor tweaks here and there, mainly to StringView. A stupid thing that
`flag?(:debug)` check was; for some reason debug is enabled in release
mode for whatever reason; so the new grammar was like a few 1000x slower
than the current one and I was briefly very very sad. Anyway, hiding the
Unicode check under a different flag, we now have decent performance
relative to the amount of flexibility the new grammar engine gives; and
also some perf bumps here and there in the tests/editor.