ml: Upload WwML progress
Basically only the dict remains now, and some test todos. How close and
yet how far away!.. The dict code is like 500 lines of pure junk to try
to refactor. And refactoring this isn't easy sensing the f**ked-upness
of all this arch. And yet what other arch do we have? I certainly won't
throw all of this into a huge grammar file, that'd be impossible to work
with... I just hope all of this will work without any problem for
several years at least. It doesn't appear this machinery can break once
you organize it well. But this organization, damn it's boring...