ml: Allow some extra characters in symbols, modify the spec slightly
Single quote `'` is now a valid character in symbols, but symbols cannot
start with it. Single quote is extremely useful for cases where there are
multiple generations of the same thing. In Crystal I'm usually numbering
them (e.g. `env0`, `env1` etc.) But in WwML I have the freedom to use
the character, `'`, which this numbering actually simulates: `env`,
`env'`, `env''`, etc. Some functional languages do this and I like
this character very much. The less names I need to come up with -- the better.

Once I'm done with UIR (I'm rather close) I'll attempt a rewrite of ML
again. Especially UIR is a huge consumer of all sorts of shorthands,
most of which we're currently missing; and a read-time templating language
which we're missing as well.